+++ gdal/ogr/ogrsf_frmts/dwg/ogrdgnv8layer.cpp
@@ -110,7 +110,7 @@ OGRDGNV8Layer::OGRDGNV8Layer( OGRDGNV8DataSource* poDS,
     if( !EQUAL(pszULinkType,"NONE") )
     {
         char * testULink;
-        iULinkType = strtol( pszULinkType, &testULink, 10 );
+        iULinkType = (int) strtol( pszULinkType, &testULink, 10 );
         if( strlen(pszULinkType) == 0 || *testULink != '\0' )
         {
             CPLError( CE_Warning, CPLE_AppDefined,
